Switch from ParsedFile tracking to IncludeBounds
Dealing with tree structures and the `FileMapping`s was a pain all in all. This is simpler and doesn't make me want to rip my hair out. A tree structure will need to be created from the bounds vector to display to the user, but this should be simple enough.