Rework file inclusion to not break on false positives
File inclusion was breaking when finding "/include/" in a linemarker resulting in the linemarker not being parsed therefore breaking the bounds tracking and a few other things along with it.

Now, the `find_include` method returns the slice leading to the include statement, the path of the file to include, and the slice after the include statement. This keeps false positives from adding part of a linemarker at a time, resulting in the linemarker not being parsed.