Create test runner
Runs the includer and parser on all dts files in "tests/dts". Tests are parallelized with rayon and are actually run under one test.

It would be nice if a seperate test function for each file could be make at compile time, but that is a task for the future.