Fix two issues with bounds creation
The first issue came about when a DTS include was at the very start of a bound. Due to old logic that hadn't been thought through after the change to bounds this did not create a zero length bound before the include which is needed to be able to figure out the include tree later on.

The second issue was that new CPP bounds sometimes did not have the proper length at creation due to only knowing of the slice of input before a DTS include. The fix just passes in the length of the remaining section. This length will be trimmed by the following split from the DTS include, but ensures that the proper remainder is left after splitting.