Allow empty cell lists and make separate Cell enum
This ballooned far past what I thought would need to be required due to funkiness with how `comments_ws!` works. In truth, the move to a dedicated Cell enum is probably not neccessary, but it does help ensure some invariants about cells, so it might as well stick around.

`Cell::Ref`s have an optional u64 for tracking phandles. I figured the field should be added now so that if phandles are ever implemented it won't be a breaking change/a pain to replace everywhere.